Quote test location argument when running tests
Both `runSingleTest` and `runTestFile` would fail if the path to the test fail contains spaces.

Note that this fix is not perfect: There are cases where simply wrapping a shell argument in quotes is not enough (for example if the argument contains quotes itself).

Fixes https://github.com/connorshea/vscode-ruby-test-adapter/issues/41